Preparing files
