+++ mock/urpa.py
@@ -1,0 +1,17 @@
+# minimal fake urpa used for monkeypatching
+
+
+def condition_factory():
+    pass
+
+
+class App:
+    pass
+
+
+class Condition:
+    pass
+
+
+class AppElement:
+    pass
