+++ requirements_dev.txt
@@ -1,0 +1,4 @@
+pylint==2.10.2
+pytest==6.2.5
+pytest-cov==2.12.1
+mypy==0.910
