+++ setup.cfg
@@ -1,0 +1,28 @@
+[metadata]
+name = urpaform
+description = Helps you fill in forms with UltimateRPA
+url = https://github.com/ultimaterpa/urpaform
+author = slavaqq
+license = MIT
+license_file = LICENSE
+platforms = win32
+
+[options]
+packages = urpaform
+install_requires =
+python_requires = >=3.7
+package_dir = =src
+zip_safe = no
+
+[options.extras_require]
+testing =
+    pylint>=2.10.2
+    pytest>=6.2.5
+    pytest-cov>=2.0
+    mypy>=0.910
+
+[options.package_data]
+urpaform = py.typed
+
+[pylint]
+--rcfile = .\.pylintrc
