+++ src/urpaform/__init__.py
@@ -1,2 +1,4 @@
+""" __init__.py """
+
 from .elements import *
 from .form import Form, FormError
