+++ tests/conftest.py
@@ -1,0 +1,3 @@
+import pytest
+
+pytest.MonkeyPatch().syspath_prepend("mock")
