+++ tests/test_urpaform.py
@@ -1,0 +1,6 @@
+
+import urpaform
+
+
+def test_equality():
+    assert 1 == 1
